requirements.txt
requests==2.31.0
opencv-python==4.10.0.84
mediapipe==0.10.18
numpy==1.26.4
Pillow==11.0.0
python-dotenv==1.0.0
screeninfo
